Hardening and cache consistency update

+++ src/pbaas/pbaas.h
@@ -789,7 +789,7 @@ public:
     std::multimap<arith_uint256, CPBaaSMergeMinedChainData *> mergeMinedTargets;
 
     LRUCache<uint160, CCurrencyDefinition> currencyDefCache;        // protected by cs_main, so doesn't need sync
-    LRUCache<std::tuple<uint160, uint32_t, bool>, CCoinbaseCurrencyState> currencyStateCache; // cached currency states @ heights + updated flag
+    LRUCache<std::tuple<uint160, uint256, bool>, CCoinbaseCurrencyState> currencyStateCache; // cached currency states @ heights + updated flag
 
     // make earned notarizations for one or more notary chains
     std::map<uint160, CNotarySystemInfo> notarySystems;
